impl updatesystrayiconstate

+++ src/xembed.rs
@@ -7,7 +7,7 @@ pub(crate) const XEMBED_EMBEDDED_NOTIFY: c_int = 0;
 pub(crate) const XEMBED_FOCUS_IN: c_int = 4;
 pub(crate) const XEMBED_MODALITY_ON: c_int = 10;
 
-pub(crate) const XEMBED_MAPPED: c_int = 1 << 0;
+pub(crate) const XEMBED_MAPPED: u64 = 1 << 0;
 pub(crate) const XEMBED_WINDOW_ACTIVATE: c_int = 1;
 pub(crate) const XEMBED_WINDOW_DEACTIVATE: c_int = 2;
